+++ static/sw.js
@@ -1,4 +1,4 @@
-const CACHE = 'kartiq-v5-3-3';
+const CACHE = 'kartiq-v5-4-0';
 
 const ASSETS = [
   '/',